Writes oneline per sequence.

+++ Strings/test_fasta_cleaner.py
@@ -8,19 +8,21 @@ def writeline(fp,text):
     fp.write('\n')
 
 class TestFastaCleaner(unittest.TestCase):
-
     def setUp(self):
         self.FILE1="test.in.fa"
         self.FILE2="test.out.fa"
         self.FILE3="test.answer.fa"
+        self.cleanup=False
     def tearDown(self):
-        try:
-            os.remove(self.FILE1)
-            os.remove(self.FILE2)
-            os.remove(self.FILE3)
-            #print("TO DO: remove the test files")
-        except Exception as e:
-            pass # This can happen even if the file is removed.
+        if (self.cleanup):
+            try:
+                os.remove(self.FILE1)
+                os.remove(self.FILE2)
+                os.remove(self.FILE3)
+            except Exception as e:
+                pass # This can happen even if the file is removed.
+        else:
+            print("Remember to remove the test files.")
 
     def test_one_good_sequence(self):
         with open(self.FILE1,'w') as fa:
@@ -39,8 +41,8 @@ class TestFastaCleaner(unittest.TestCase):
             writeline(fa,sequence)
         with open(self.FILE3,'w') as fa:
             writeline(fa,">upper_and_lower")
-            writeline(fa,sequence.upper())
-            writeline(fa,sequence.upper())
+            concat = sequence+sequence
+            writeline(fa,concat.upper())
         fc = Fasta_Cleaner(self.FILE1,self.FILE2)
         fc.fix_everything()
         same=os.system("diff %s %s"%(self.FILE2,self.FILE3))
